📝 [KMP] Update readme, add script for renaming project, move service to data, update konsist tests
