+++ shared/samplefeature/src/commonMain/kotlin/kmp/shared/samplefeature/data/source/impl/JokeSourceImpl.kt
@@ -3,7 +3,7 @@ package kmp.shared.samplefeature.data.source.impl
 import kmp.shared.base.domain.model.Result
 import kmp.shared.samplefeature.data.model.JokeDto
 import kmp.shared.samplefeature.data.source.JokeSource
-import kmp.shared.samplefeature.infrastructure.service.JokeService
+import kmp.shared.samplefeature.data.service.JokeService
 
 internal class JokeSourceImpl(private val service: JokeService) : JokeSource {
