+++ shared/samplefeature/src/commonMain/kotlin/kmp/shared/samplefeature/di/Module.kt
@@ -6,7 +6,7 @@ import kmp.shared.samplefeature.data.source.impl.JokeSourceImpl
 import kmp.shared.samplefeature.domain.repository.JokeRepository
 import kmp.shared.samplefeature.domain.usecase.GetRandomJokeUseCase
 import kmp.shared.samplefeature.domain.usecase.GetRandomJokeUseCaseImpl
-import kmp.shared.samplefeature.infrastructure.service.JokeService
+import kmp.shared.samplefeature.data.service.JokeService
 import kmp.shared.samplefeature.presentation.vm.SampleFeatureViewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.singleOf
