+++ shared/umbrella/src/androidUnitTest/kotlin/konsistTest/data/DataTest.kt
@@ -1,4 +1,4 @@
-package konsistTest.infrastructure
+package konsistTest.data
 
 import com.lemonappdev.konsist.api.Konsist
 import com.lemonappdev.konsist.api.ext.list.print
@@ -10,35 +10,35 @@ import com.lemonappdev.konsist.api.verify.assertTrue
 import kotlinx.serialization.Serializable
 import org.junit.Test
 
-internal class InfrastructureTest {
+internal class DataTest {
     @Test
     fun `all DTOs are internal or private`() {
         Konsist
             .scopeFromProject()
             .classes()
-            .withPackage("..infrastructure.dto..")
+            .withPackage("..data.model..")
             .assertTrue { klass ->
                 klass.hasInternalModifier || klass.hasPrivateModifier
             }
     }
 
     @Test
-    fun `all classes in 'infrastructure' package are annotated with Serializable`() {
+    fun `all classes in 'data model' package are annotated with Serializable`() {
         Konsist
             .scopeFromProject()
             .classes()
-            .withPackage("..infrastructure.dto..")
+            .withPackage("..data.model..")
             .assertTrue { klass ->
                 klass.hasAnnotationOf(Serializable::class)
             }
     }
 
     @Test
-    fun `all classes in 'infrastructure' have immutable properties`() {
+    fun `all classes in 'data model' have immutable properties`() {
         Konsist
             .scopeFromProject()
             .classes()
-            .withPackage("..infrastructure.dto..")
+            .withPackage("..data.model..")
             .assertTrue { klass ->
                 klass.properties()
                     .all { prop -> prop.hasValModifier }
@@ -50,20 +50,20 @@ internal class InfrastructureTest {
         Konsist
             .scopeFromProject()
             .functions()
-            .withPackage("..infrastructure..", "..data..")
+            .withPackage("..data..")
             .withName("toDomain")
             .print { it.fullyQualifiedName }
             .assertTrue { fn -> fn.hasInternalModifier || fn.hasPrivateModifier }
     }
 
     @Test
-    fun `all 'toDomain()' function are in the 'infrastructure' package`() {
+    fun `all 'toDomain()' function are in the 'data' package`() {
         Konsist
             .scopeFromProject()
             .functions()
             .filter { it.isTopLevel }
             .withoutReceiverType { it.name.contains("ViewObject") }
             .withNameContaining("toDomain")
-            .assertTrue { it.resideInPackage("..infrastructure..") }
+            .assertTrue { it.resideInPackage("..data..") }
     }
 }
